test: move upstream tla deps from testlib to its file
